+++ app/services/payment_service.rb
@@ -16,19 +16,19 @@ class PaymentService < ApplicationService
   def deduct_credit(payment)
     return true if CustomerCredit.deduct!(payment)
 
-    Sentry.capture_exception('Unable to deducet from gift card', transaction: 'Post Payment')
+    Sentry.capture_exception('Unable to deduct from gift card', transaction: 'Post Payment')
     raise ProcessError, 'Unable to deduct from gift card'
   end
 
   def post_to_braintree(payment)
-    result = gateway.post_sale payment
+    result = braintree.post_sale payment
     payment.identifier = result.transaction.id
     payment.method = result.transaction.payment_instrument_type
 
     true
   end
 
-  def gateway
-    @gateway ||= BraintreeService.new
+  def braintree
+    @braintree ||= BraintreeService.new
   end
 end